Mapper de persona
Se agrega el mapper de persona para la creación de sus accesos a bases de datos desde la capa de aplicación.

+++ PackageDelivery.Repository,Contracts/DbModels/Parameters/PersonDbModel.cs
@@ -14,7 +14,7 @@
         public string cellphone { get; set; }
         public string email { get; set; }
 
-        public long IdAddress { get; set; }
+        public int IdAddress { get; set; }
         public int IdDocumentType { get; set; }
 
     }
